[4301] Fix up issues introduced by merge conflicts
